scripts/requirements.txt
# AI Code Review Agent dependencies
# Install only the provider you use — all are optional
anthropic>=0.39.0    # for Anthropic Claude
openai>=1.50.0       # for OpenAI GPT
google-generativeai>=0.8.0  # for Google Gemini
